поддержка различных значимых типов через switch(type.Name)

+++ Solution/Test/Program.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
+using System.Runtime.Serialization.Formatters.Binary;
 using System.Text;
 using System.Collections;
 using SecureField;
@@ -11,8 +13,16 @@ namespace Test
     {
         static void Main(string[] args)
         {
+            //var bitv = new BitValue(new object());
+            //var mask = new byte[] {255};
+            //bitv.Bytes = bitv.Bytes.Xor(mask);
+            //var newVal = bitv.ToValue();
+            //bitv = new BitValue(newVal);
+            //bitv.Bytes = bitv.Bytes.Xor(mask);
+            //var oldVal = bitv.ToValue();
+
+
             var val = new Values();
-            //var metdos = typeof(Values).GetProperty("Number").GetSetMethod();
             while (true) 
             {
                 Console.Write("Enter: ");
@@ -24,7 +34,7 @@ namespace Test
                 }
                 else
                 {
-                    var n = int.Parse(str);
+                    var n = float.Parse(str);
                     val.Number = n;
                 }
